requirement.txt
nextcord
Pillow
pyzbar
